data_ingestion/structured_data/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-socketio
eventlet
pydantic
yfinance
pandas
requests
feedparser
python-dotenv
python-multipart
